chore: add images to fake markdown
